+++ src/fluffycontrol/RestrictionManager.java
@@ -16,7 +16,7 @@ public class RestrictionManager {
             // Nighttime. After 23:30 and before 8:00
             if (!isDayTime && !fired[0]) {
                 System.out.println("Get to bed");
-                warningWindow("Sleep time","You can do whatever you need to do tomorrow.","#D0342C");
+                SwingUtilities.invokeLater(warnWindow::new);
                 // Turn off Ethernet
                 NetworkController.turnOffNetwork();
                 fired[0] = true; // Lock night event
@@ -37,29 +37,5 @@ public class RestrictionManager {
     // TimerCheck Helper
     private static boolean isBetween(LocalTime now, LocalTime start, LocalTime end) {
         return !now.isBefore(start) && !now.isAfter(end);
-    }
-
-    public static void warningWindow(String title, String body, String color){
-        JFrame frame = new JFrame();
-
-        //frame.setUndecorated(true);
-        frame.setAlwaysOnTop(true);
-        frame.setSize(850,600);
-        frame.setLocationRelativeTo(null);
-        //frame.setExtendedState(JFrame.MAXIMIZED_BOTH);
-        frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-
-        JLabel label = new JLabel(
-                "<html><center><h1>" + title + "</h1><br>" + body + "</center></html>",
-                SwingConstants.CENTER
-        );
-
-        label.setFont(new Font("Verdana", Font.BOLD, 40));
-        label.setBackground(Color.decode(color));
-
-        label.setOpaque(true);
-
-        frame.add(label);
-        frame.setVisible(true);
     }
 }